# client-operator/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Plasma-Client implementation",
        "title": "Swagger History API",
        "contact": {
            "name": "API Support",
            "email": "nk@bankexfoundation.org, kk@bankexfoundation.org, ig@bankexfoundation.org, av@bankexfoundation.org"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.0.1"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/SetMyTx": {
            "post": {
                "description": "Put your tx by tx Id",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Put your tx by tx Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tx",
                        "name": "tx",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.SuccessResponse"
                            }
                        }
                    }
                }
            }
        },
        "/getHistoryPart": {
            "get": {
                "description": "Return History of blocks",
                "produces": [
                    "application/json"
                ],
                "summary": "Return History of blocks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.HistoryResponse"
                            }
                        }
                    }
                }
            }
        },
        "/getMyTx": {
            "get": {
                "description": "Return Tx by tx Id",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Return Tx by tx Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tx",
                        "name": "tx",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.TxResponse"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "responses.HistoryResponse": {
            "type": "object",
            "properties": {
                "History": {
                    "type": "string",
                    "example": "1c8d54df80c03a56b5470d164c49f823108f96a67d020e4c677810c9a10b1ca7"
                }
            }
        },
        "responses.SuccessResponse": {
            "type": "object",
            "properties": {
                "Status": {
                    "type": "string",
                    "example": "Ok"
                }
            }
        },
        "responses.TxResponse": {
            "type": "object",
            "properties": {
                "Tx": {
                    "type": "string"
                }
            }
        }
    }
}